feat: Corrects the LazyLoadEntryPointContainer_DEPRECATED stuffs
